move 3d annotation coercePosition step from convert to defaults
- as now Axes.coercePosition does not rely on
  ax._categories for `type: 'category' axes.
